+add access non Admin users for get/update/delete self account +add Access_is_denied for ExceptionHandler

+++ src/main/java/ru/dsstaroverov/lunchMenu/web/UserRestController.java
@@ -7,6 +7,7 @@ import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.access.AccessDeniedException;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.servlet.support.ServletUriComponentsBuilder;
 import ru.dsstaroverov.lunchMenu.model.User;
@@ -18,6 +19,7 @@ import java.net.URI;
 import java.util.List;
 
 import static ru.dsstaroverov.lunchMenu.util.ValidationUtil.assureIdConsistent;
+import static ru.dsstaroverov.lunchMenu.web.SecurityUtil.*;
 
 @RestController
 @RequestMapping(value = UserRestController.REST_URL, produces = MediaType.APPLICATION_JSON_VALUE)
@@ -33,29 +35,51 @@ public class UserRestController {
 
     @GetMapping
     public List<User> getAll() {
-        return userService.getAll();
+        if (isAdmin()) {
+            log.info("getAll users");
+            return userService.getAll();
+        }
+        throw new AccessDeniedException("");
     }
 
 
     @GetMapping("/{id}")
     public User get(@PathVariable int id) {
-        return userService.get(id);
+        if (isAdmin()) {
+            log.info("get with id={}", id);
+            return userService.get(id);
+        }else {
+            log.info("get with id={}", authUserId());
+            return userService.get(authUserId());
+        }
+
     }
 
     @PostMapping(consumes = MediaType.APPLICATION_JSON_VALUE)
     public ResponseEntity<User> create(@Valid @RequestBody User user) {
-        User created = userService.create(user);
-        URI uriOfNewResource = ServletUriComponentsBuilder.fromCurrentContextPath()
-                .path(REST_URL + "/{id}")
-                .buildAndExpand(created.getId()).toUri();
-        return ResponseEntity.created(uriOfNewResource).body(created);
+        if (isAdmin()) {
+            log.info("create user");
+            User created = userService.create(user);
+            URI uriOfNewResource = ServletUriComponentsBuilder.fromCurrentContextPath()
+                    .path(REST_URL + "/{id}")
+                    .buildAndExpand(created.getId()).toUri();
+            return ResponseEntity.created(uriOfNewResource).body(created);
+        }
+        throw new AccessDeniedException("");
     }
 
 
     @DeleteMapping("/{id}")
     @ResponseStatus(HttpStatus.NO_CONTENT)
     public void delete(@PathVariable int id) {
-        userService.delete(id);
+        if (isAdmin()) {
+            log.info("delete user with id={}", id);
+            userService.delete(id);
+        }else {
+            SecurityUtil.get().getUser().setEnabled(false);
+            log.info("user with id={} set enabled false", authUserId());
+            userService.update(SecurityUtil.get().getUser());
+        }
     }
 
 
@@ -63,9 +87,15 @@ public class UserRestController {
     @ResponseStatus(value = HttpStatus.NO_CONTENT)
     public void update(@Valid @RequestBody User user, @PathVariable int id) {
         try {
-            log.info("update {} with id={}", user, id);
-            assureIdConsistent(user, id);
-            userService.update(user);
+            if (isAdmin()) {
+                log.info("update {} with id={}", user, id);
+                assureIdConsistent(user, id);
+                userService.update(user);
+            }else {
+                assureIdConsistent(user, authUserId());
+                log.info("update {} with id={}", user, authUserId());
+                userService.update(user);
+            }
         }catch (DataIntegrityViolationException e){
             throw new IllegalRequestDataException("email dublicate update");
         }
@@ -74,6 +104,9 @@ public class UserRestController {
 
     @GetMapping("/by")
     public User getByMail(@RequestParam String email) {
-        return userService.getByEmail(email);
+        if (isAdmin()) {
+            return userService.getByEmail(email);
+        }
+        throw new AccessDeniedException("");
     }
 }
